+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * new module: pulse ([#223][223]).
 * alsa: `dB` tag ([#202][202]).
 * mpd: `file` tag ([#219][219]).
+* pipewire: add a new module for pipewire ([#224][224])
 * on-click: support `next`/`previous` mouse buttons ([#228][228]).
 
 [153]: https://codeberg.org/dnkl/yambar/issues/153
@@ -38,6 +39,7 @@
 [202]: https://codeberg.org/dnkl/yambar/issues/202
 [219]: https://codeberg.org/dnkl/yambar/pulls/219
 [223]: https://codeberg.org/dnkl/yambar/pulls/223
+[224]: https://codeberg.org/dnkl/yambar/pulls/224
 [228]: https://codeberg.org/dnkl/yambar/pulls/228